requirements.txt
asynctest==0.13.0
fastapi==0.95.0
httpx==0.23.3
nltk==3.8.1
pytest-asyncio==0.21.0
uvicorn==0.21.1
